Resume Screening & Job Matching Assistant/requirements.txt
openai
streamlit
pandas
PyMuPDF
python-docx
Pillow
pytesseract
python-dotenv
